+++ src/common/mod.rs
@@ -1,4 +1,4 @@
-#[cfg(all(feature = "bpf", target_os = "linux"))]
+#[cfg(target_os = "linux")]
 pub mod bpf;
 
 pub mod classic;