+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Changed
 
+- BPF is now non-optional on Linux. (#336)
 - Makes HTTP compression dependent on client accept headers, removes the need
   for separate configuration of compression. (#337)